gateway-service/requirements.txt
Django
djangorestframework
python-dotenv
requests
gunicorn
coverage
flake8
black
